recap_balance/requirements.txt
ccxt
datetime
configparser
mysql-connector-python



